+++ JAM mkII/Areas/Admin/Views/User/Add.cshtml
@@ -1,13 +1,13 @@
-﻿@model RegisterViewModel
+﻿@model JAM_mkII.Models.ViewModels.RegisterViewModel
 @{
-    ViewBag.Title = "Edit User ";
+    ViewBag.Title = "Add User";
 }
 
-<h2>Edit User: <span>@Model.Email</span></h2>
+<h2>Add User</h2>
 
 <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-<form method="post" asp-action="Edit">
-@*   <div class="form-group row">
+<form method="post" asp-action="Add">
+   <div class="form-group row">
         <div class="col-sm-2"><label>Email address</label></div>
         <div class="col-sm-4">
             <input asp-for="Email" class="form-control"/>
@@ -15,7 +15,7 @@
         <div class="col">
             <span asp-validation-for="Email"
                   class="text-danger"></span>
-        </div>*@
+        </div>
     </div>
     <div class="form-group row">
         <div class="col-sm-2"><label>First Name:</label></div>
