+++ JAM mkII/Areas/Admin/Views/User/Index.cshtml
@@ -1,4 +1,5 @@
-﻿@model UserViewModel
+﻿@using JAM_mkII.Models.DomainModels
+@model JAM_mkII.Areas.Admin.Models.UserViewModel
 @{
     ViewData["Title"] = " | Manage Users";
 }
