+++ JAM mkII/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using JAM_mkII
+@using JAM_mkII.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
