+++ JAM mkII/Models/DomainModels/Employment.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace JAM_mkII.Models
+namespace JAM_mkII.Models.DomainModels
 {
     public class Employment
     {
