+++ JAM mkII/Models/DomainModels/Result.cs
@@ -1,4 +1,4 @@
-﻿namespace JAM_mkII.Models
+﻿namespace JAM_mkII.Models.DomainModels
 {
     public class Result
     {
