+++ JAM mkII/Models/ViewModels/ChangePasswordViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
-namespace JAM_mkII.Models
+
+namespace JAM_mkII.Models.ViewModels
 {
     public class ChangePasswordViewModel
     {
