+++ JAM mkII/Models/ViewModels/LoginViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace JAM_mkII.Models
+namespace JAM_mkII.Models.ViewModels
 {
     public class LoginViewModel
     {
