+++ JAM mkII/Models/ViewModels/RegisterViewModel.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace JAM_mkII.Models
+namespace JAM_mkII.Models.ViewModels
 {
-    public class EditViewModel
+    public class RegisterViewModel
     {
         [Required(ErrorMessage = "Please enter valid email address.")]
         [DataType(DataType.EmailAddress)]
